Added safeties without json to resolve #223.
Free Software Free Society

To support commits by ctubio,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/server/og.h
@@ -230,7 +230,7 @@ namespace K {
           {"tradeId", k.tradeId},
           {"time", k.time},
           {"exchange", (int)k.exchange},
-          {"pair", {{"base", (int)k.pair.base}, {"quote", (int)k.pair.quote}}},
+          {"pair", {{"base", k.pair.base}, {"quote", k.pair.quote}}},
           {"price", k.price},
           {"quantity", k.quantity},
           {"side", (int)k.side},
